+++ src/ebird/checklists/models/__init__.py
@@ -1,7 +1,8 @@
 from .checklist import Checklist
+from .country import Country
 from .location import Location
 from .observation import Observation
 from .observer import Observer
 from .species import Species
 
-__all__ = ("Checklist", "Location", "Observation", "Observer", "Species")
+__all__ = ("Checklist", "Country", "Location", "Observation", "Observer", "Species")