+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0
 Latest
 ------
 * Add date to Observation to simplify queries.
+* Create separate Country model from Location.
 
 1.0.3 (2025-03-20)
 ------------------